+++ utoipa-gen/CHANGELOG.md
@@ -1,16 +1,19 @@
 # Changelog - utoipa-gen
 
-## Unreleased
+## 5.1.2 - Oct 16 2024
 
 ### Added
 
 * Add implementation for utoipa-actix-web bindings (https://github.com/juhaku/utoipa/pull/1158)
 
+### Changed
+
+* Finalize actix-web utoipa bindings (https://github.com/juhaku/utoipa/pull/1160)
+
 ## 5.1.1 - Oct 16 2024
 
 ### Changed
 
-* Finalize actix-web utoipa bindings (https://github.com/juhaku/utoipa/pull/1160)
 *  Enhance no_recursion rule to apply also containers (https://github.com/juhaku/utoipa/pull/1144)
 
 ## 5.1.0 - Oct 16 2024